Fourth Commit
This is my last commit for the afternoon. Karl, if you could prioritize the "trace out" and "measure" GUI's it would be great as they are required to represent the bb84 list. The toString method is just about done otherwise, so we should have something neat to show her tomorrow.